Bug Fix: Implement neumorphism shadow set retry functionality
In Ionic (with Web Components), the DOM was not fully loaded by the time ngAfterViewChecked was running.
This hooks into a change detection lifecycle methods, opposed to an initializing method and retries when the DOM is loaded (when element height is greater than 0).

Optimization concerns: The shadow logic has an average case of 1 run and worst case of 2 runs because it does not proceed if the shadow logic was ever successful.